Added "colNorm" and renamed "norm" to "rowNorm" to clarify it. Updated norm dependencies to reflect move from LinAlg to Stats. Expanded test to separately test row- and column-wise norms
